refactor: added a save method, which saves computer values
