+++ Program.cs
@@ -27,8 +27,7 @@ dotnet add package Microsoft.Data.Sqlite -s 'C:\Users\IFSP\.nuget\packages'
 using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
-
-
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
 
@@ -58,21 +57,18 @@ if (modelName == "Computer")
         var id = Convert.ToInt32(args[2]);
         var ram = args[3];
         var processor = args[4];
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-        var command = connection.CreateCommand();
 
-        command.CommandText = $"INSERT INTO Computers VALUES($id, $ram, $processor)";
+        var computer = new Computer(id, ram, processor);
 
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-
-        command.ExecuteNonQuery();
-
-        connection.Close();
+        computerRepository.Save(computer);
+        
+        
     }
 }
+
+
+
+
 else if (modelName == "Lab")
 {
     switch (modelAction)
