+++ Repositories/ComputerRepository.cs
@@ -47,4 +47,28 @@ class ComputerRepository
 
         return computers;
     }
+
+    public void Save(Computer computer)
+    {   
+        var connection = new SqliteConnection(_databaseConfig.ConnectionString);
+        connection.Open();
+        var command = connection.CreateCommand();
+
+        command.CommandText = $"INSERT INTO Computers VALUES($id, $ram, $processor)";
+
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Ram);
+
+        command.ExecuteNonQuery();
+
+        connection.Close();
+    }
+
+    /*
+    public void Save(int id, string ram, string processor)
+    {
+
+    }
+    */
 }
